Readd Notes feature
The Script now allows you to read the config from your notes and therefore ensure cross-device saves. Sadly, the way I tried to save to the notes seems to have a character limit, as seen in [#6]. Saving to the notes is disabled as a result, and the script will show you the raw data for you to manually paste into the notes.